Updeta code base from reviews
