--- utils/fork_star_desc.js
@@ -1,10 +1,0 @@
-const axios = require('axios')
-
-async function getForkStarDesc (user, repository) {
-  const url = await axios.get(
-    `https://api.github.com/repos/${user}/${repository}`
-  )
-  return url.data
-}
-
-module.exports = getForkStarDesc
