+++ utils/index.js
@@ -1,5 +1,5 @@
 const getIssues = require('./issue')
-const getForkStarDesc = require('./fork_star_desc')
+const getForkStarDesc = require('./fork-star-desc')
 
 module.exports = {
   getIssues,
